Better sharing qr address screen
