+++ app/src/main/java/com/vultisig/wallet/ui/components/QRCodeKeyGenImage.kt
@@ -1,7 +1,6 @@
 package com.vultisig.wallet.ui.components
 
 import android.graphics.Bitmap
-import android.graphics.Canvas
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -23,29 +22,23 @@ import androidx.compose.ui.graphics.PathEffect
 import androidx.compose.ui.graphics.asImageBitmap
 import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.graphics.painter.BitmapPainter
-import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import androidx.core.graphics.scale
 import androidx.core.graphics.toColorInt
-import com.google.zxing.BarcodeFormat
-import com.google.zxing.EncodeHintType
 import com.google.zxing.WriterException
-import com.google.zxing.qrcode.QRCodeWriter
 import com.vultisig.wallet.ui.theme.Theme
+import com.vultisig.wallet.ui.utils.generateQrBitmap
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlin.math.PI
-
-private const val QR_CODE_SCALE_FACTOR = 8
-private const val QR_CODE_VS_LOGO_SCALE_FACTOR = 4
 
 @Composable
 internal fun QRCodeKeyGenImage(
     qrCodeContent: String,
     modifier: Modifier = Modifier,
+    bitmapPainter: BitmapPainter? = null,
     cornerRadius: Dp = 16.dp,
     innerPadding: Dp = 24.dp,
     relativePaintedDashLength: Float = 7f,
@@ -80,7 +73,7 @@ internal fun QRCodeKeyGenImage(
                 .padding(innerPadding),
         ) {
             Image(
-                painter = rememberQRBitmapPainter(
+                painter = bitmapPainter ?: rememberQRBitmapPainter(
                     qrCodeContent = qrCodeContent,
                 ),
                 contentDescription = "qr",
@@ -131,65 +124,7 @@ internal fun rememberQRBitmapPainter(
         )
     }
 }
- internal fun generateQrBitmap(
-    qrCodeContent: String,
-    mainColor: Color = Color.Black,
-    backgroundColor: Color = Color.White,
-    logo : Bitmap? = null,
-): Bitmap {
-    val hintMap = mapOf(EncodeHintType.MARGIN to 0)
 
-    val qrCodeWriter = QRCodeWriter()
-    val bitmapMatrix = qrCodeWriter.encode(
-        qrCodeContent,
-        BarcodeFormat.QR_CODE,
-        0,
-        0,
-        hintMap
-    )
-
-    val matrixWidth = bitmapMatrix.width
-    val matrixHeight = bitmapMatrix.height
-
-    val bitmap = Bitmap.createBitmap(
-        matrixWidth,
-        matrixHeight,
-        Bitmap.Config.ARGB_8888,
-    )
-
-    for (x in 0 until matrixWidth) {
-        for (y in 0 until matrixHeight) {
-            val shouldColorPixel = bitmapMatrix?.get(x, y) ?: false
-            val pixelColor =
-                if (shouldColorPixel) mainColor.toArgb() else backgroundColor.toArgb()
-
-            bitmap.setPixel(x, y, pixelColor)
-        }
-    }
-    if (logo == null) return bitmap
-
-    val scaledWidth = bitmap.width * QR_CODE_SCALE_FACTOR
-    val scaledHeight = bitmap.height * QR_CODE_SCALE_FACTOR
-    val scaledLogoWidthTemp = scaledWidth / QR_CODE_VS_LOGO_SCALE_FACTOR
-    val scaledLogoHeightTemp = scaledHeight / QR_CODE_VS_LOGO_SCALE_FACTOR
-    val scaledLogoWidth = if (scaledLogoWidthTemp == 0) 1 else scaledLogoWidthTemp
-    val scaledLogoHeight = if (scaledLogoHeightTemp == 0) 1 else scaledLogoHeightTemp
-    val scaledLogo = logo.scale(scaledLogoWidth, scaledLogoHeight)
-    val scaledBitmap = bitmap.scale(scaledWidth, scaledHeight, false)
-
-    val canvas = Canvas(scaledBitmap)
-
-    val xLogo = (scaledBitmap.width - scaledLogo.width) / 2f
-    val yLogo = (scaledBitmap.height - scaledLogo.height) / 2f
-
-    canvas.drawBitmap(scaledLogo, xLogo, yLogo, null)
-    if (bitmap != scaledBitmap)
-        bitmap.recycle()
-
-    scaledLogo.recycle()
-
-    return scaledBitmap
-}
 
 @Preview
 @Composable
