+++ app/src/main/java/com/vultisig/wallet/ui/models/swap/SwapViewModel.kt
@@ -9,7 +9,6 @@ import androidx.lifecycle.viewModelScope
 import com.vultisig.wallet.data.models.Vault
 import com.vultisig.wallet.data.repositories.VaultRepository
 import com.vultisig.wallet.ui.utils.ShareType
-import com.vultisig.wallet.ui.components.generateQrBitmap
 import com.vultisig.wallet.ui.utils.share
 import com.vultisig.wallet.ui.utils.shareFileName
 import com.vultisig.wallet.ui.models.AddressProvider
@@ -17,6 +16,7 @@ import com.vultisig.wallet.ui.navigation.Destination
 import com.vultisig.wallet.ui.navigation.NavigationOptions
 import com.vultisig.wallet.ui.navigation.Navigator
 import com.vultisig.wallet.ui.navigation.SendDst
+import com.vultisig.wallet.ui.utils.generateQrBitmap
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
@@ -60,7 +60,7 @@ internal class SwapViewModel @Inject constructor(
         }
     }
 
-    internal fun shareQRCode(activity: Context): Unit {
+    internal fun shareQRCode(activity: Context) {
         val qrBitmap = generateQrBitmap(addressProvider.address.value)
         activity.share(
             qrBitmap,
